requirements.txt
psycopg[binary]
flask
bcrypt
pyjwt
python-dotenv